move PostConfLoadedSetup to init.c
