+++ src/tests/fuzz/fuzz_applayerparserparse.c
@@ -12,6 +12,8 @@
 
 #define HEADER_LEN 6
 
+int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size);
+
 AppLayerParserThreadCtx *alp_tctx = NULL;
 
 int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
